requirements.txt
asgiref==3.8.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
Django==5.2.1
django-easy-audit==1.3.7
django-maintenance-mode==0.22.0
django-unfold==0.58.0
dotenv==0.9.9
et_xmlfile==2.0.0
firebase-admin==6.8.0
gcloud==0.18.3
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
lxml==5.4.0
msgpack==1.1.0
oauth2client==4.1.3
openpyxl==3.1.5
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.3
Pyrebase4==4.8.0
python-docx==1.1.2
python-dotenv==1.1.0
python-fsutil==0.15.0
python-jwt==4.1.0
qrcode==8.2
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.26.20
whitenoise==6.9.0
